feat: blog redesign vol.3 (#111)
* initial blog redesign

* color palet and element position changes

* rating display redesign

* single blog checkpoint

checkpoint

glavni

* started blog list display

* blog display redesign

* checkpoint

* blog-form redesign

* kk

* chcek

* checkpoint

* color change

* a

* redesigned

* checkpoint

* logic fix

* code refactor

* comments display redesign

* added image display on blog page

* comment body change

* color change

* fml

---------

+++ Explorer/src/app/feature-modules/blog/comments-display/comments-display.component.ts
@@ -2,6 +2,7 @@ import { Component,Input,Output,OnInit,EventEmitter } from '@angular/core';
 import { CommentService } from '../comment.service';
 import { Comment } from './../model/comment.model';
 import { PagedResults } from 'src/app/shared/model/paged-results.model';
+import { Blog } from '../model/blog.model';
 
 @Component({
   selector: 'xp-comments-display',
@@ -10,33 +11,30 @@ import { PagedResults } from 'src/app/shared/model/paged-results.model';
 })
 export class CommentsDisplayComponent implements OnInit {
   
-  @Input() blogId : number
+  @Input() blog : Blog
   @Input() systemStatus : string
   
 public editMode : boolean
 public comment : Comment
+public userId : number
 
   public comments: Comment[] = []
 
   constructor(private commentService: CommentService){}
 
   ngOnInit(): void{
-
-    this.commentService.getComments(0,0,this.blogId).subscribe({
-      next: (result: PagedResults<Comment>) =>{
-        this.comments = result.results;
-      },
-      error: (err: any) => {
-        console.log(err);
-      },
-      
-    })
+    this.getComments();
+    this.userId = parseInt(localStorage.getItem('loggedId')??'1');
   }
 
   getComments(): void{
-    this.commentService.getComments(0,0,this.blogId).subscribe({
+    if(this.blog.id === undefined){
+      return
+    }
+    this.commentService.getComments(0,0,this.blog.id).subscribe({
       next : (response: PagedResults<Comment>)=>{
         this.comments = response.results;
+        this.comments.reverse();
         this.editMode = false;
       },
       error:(err : any)=>{
@@ -47,18 +45,22 @@ public comment : Comment
 
 onUpdateClicked(comment: Comment): void
 {
+  if(this.userId !== comment.userId || this.blog.systemStatus === 'CLOSED'){
+    return
+  }
   this.editMode = true;
   this.comment = comment;
 }
 
 onDeleteClicked(comment: Comment): void
 {
+  if(this.userId !== comment.userId || this.blog.systemStatus === 'CLOSED'){
+    return
+  }
   this.commentService.deleteComment(comment).subscribe({
     next: (_) => {
       this.getComments();
     }
   });
 }
-
-
 }
